Add hand scan sensation and set it as the default sensation
The method `setScanEnabled` is intended to be called externally to
change the sensation type between hand scan and circle.

The position the hand needs to be in for the hand scan is defined by the
position constraints in `LeapListening::onFrame`.

The default hand scan parameters are set in the `HandScan` struct. E.g.
1.5 second vertical scan duration as the default.